bugfix: do not re-ask for currentItem.asset.duration
* we already have _videoDuration as an instance variable
* when asking the player for asset duration the first time, it can take
up to 350ms to return in certain scenarios. Specifically, when using
AVPlayerLayer (as opposed to AVPlayerViewController).
* Luckily, we already get the _videoDuration above, and we smartly
wrap it in a check to see if the duration is loaded

```
AVKeyValueStatus status = [_player.currentItem.asset statusOfValueForKey:@"duration" error: &error];
if (status == AVKeyValueStatusLoaded) {
  CMTime duration = [_player.currentItem.asset duration];
  if (CMTimeCompare(_videoDuration, duration) != 0) {
    _videoDuration = duration;
  }
}
```

In this conditional we will use the _videoDuration that we already
have set instead of re-requesting it.